labs/old/abitur/pyproject.toml
[tool.black]
line-length = 79
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  | migrations
  | settings
)/
'''

[tool.poetry]
name = "web"
version = "0.1.0"
description = "Web-application for lk-abit project (web enrollment application in MAI)"
authors = ["S. Gulyaev <gulyaevsa@mai.ru>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.1.7"
python-dotenv = "^0.16.0"
django-debug-toolbar = "^3.2"
daphne = "^3.0.1"
requests = "^2.25.1"
SeleniumLogin = "^0.0.4"
django-selenium-login = "^2.0.0"
behave-django = "^1.4.0"
psycopg2 = "^2.8.6"
zeep = "^4.0.0"
urllib3 = "^1.26.4"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
isort = "^5.8.0"
mypy = "^0.812"
pre-commit = "^2.11.1"
pytest = "^6.2.2"
pytest-django = "^4.1.0"
pytest-cov = "^2.11.1"
pylint = "^2.7.2"
selenium = "^3.141.0"
pylint-django = "^2.4.2"
factory-boy = "^3.2.0"
behave = "^1.2.6"
behave-django = "^1.4.0"
bandit = "^1.7.0"
testcontainers = "^3.4.0"
django-debug-toolbar = "^3.2"
geckodriver-autoinstaller = "^0.1.0"
pluggy = "^0.13.1"
urllib3 = "^1.26.4"

[tool.pylint.master]

max-line-length = 78
ignore = [
    "settings",
    "migrations"
]


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
